% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baked_goods}
\alias{baked_goods}
\title{Cupcake and Muffin Ingredients}
\format{
\subsection{\code{baked_goods}}{

A data frame with 870 rows and 4 columns:
\describe{
\item{\code{type}}{muffin or cupcake}
\item{\code{recipe_id}}{unique recipe ID}
\item{\code{ingredient}}{name of ingredient}
\item{\code{ammount_cups}}{ammount of ingredient in US cups}
}
}
}
\source{
\url{https://allrecipes.com}
}
\usage{
baked_goods
}
\description{
Ingredients for cupcake and muffin recipes from allrecipes.com
}
\keyword{datasets}
