% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bar_plot_data}
\alias{bar_plot_data}
\title{Barplot function}
\usage{
bar_plot_data(data, value, startdate = data$Date[1], rollm = FALSE, title)
}
\arguments{
\item{data}{\code{data.frame}, covid19 data.}

\item{value}{\code{character}, Column name variable of data.}

\item{startdate}{\code{character} or \code{Date}, format: "\%Y-\%m-\%d", default start}

\item{rollm}{\code{logical}, if TRUE then a line with rolled weekly mean will be added}

\item{title}{\code{character}, Title text, if missing no title}
}
\value{
data.frame
}
\description{
Barplot function
}
\details{
it plots value over Date
}
\examples{
library(ggplot2)
df <- read_data("case_time_series.csv","covid_data")
data <- process_data(df)
bar_plot_data(data, "Daily.Confirmed", rollm = TRUE)
bar_plot_data(data, "Daily.Active")
}
