\name{convolveR}
\alias{convolveR}
\title{Convolution of two vector}
\usage{
  convolveR(x, y)
}
\arguments{
  \item{x}{numeric vector}

  \item{y}{numeric vector}
}
\value{
  numeric vector of length (nx+ny-1)
}
\description{
  A description of this function
}
\details{
  Details about this function
}
\author{
  weicheng
}

