% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_exp}
\alias{df_exp}
\title{Data frame of exponential data}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{x}{1 to 10}
  \item{y}{(1 to 10)^2}
  ...
}
}
\source{
\url{https://noaa-iea.github.io/r3-train/packages.html}
}
\usage{
df_exp
}
\description{
A simple data frame of y ~ x^2
}
\keyword{datasets}
