% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.ColorBars}
\alias{.ColorBars}
\title{Color for bar plot}
\format{
An object of class \code{character} of length 1.
}
\usage{
.ColorBars
}
\description{
Color for bar plot
}
\keyword{datasets}
