% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.ColorLines}
\alias{.ColorLines}
\title{Color for lines in line plot}
\format{
An object of class \code{character} of length 1.
}
\usage{
.ColorLines
}
\description{
Color for lines in line plot
}
\keyword{datasets}
