% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dumbfunc.R
\name{dumbfunc}
\alias{dumbfunc}
\title{A dumb function}
\usage{
dumbfunc(x, y)
}
\arguments{
\item{x}{The transformed data matrix with the variables in rows and samples in columns}

\item{y}{The model matrix being used to fit the data}
}
\value{
sum The sum of the values
}
\description{
This dumb function does some dumb stuff like add two numbers. See also
}

