% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOptions_Deterministic.R
\name{EPutOption_pricingfunction_det}
\alias{EPutOption_pricingfunction_det}
\title{EPutOption_pricingfunction_det
Give the price of an Asian Put Option - Deterministic Formula p. 46}
\usage{
EPutOption_pricingfunction_det(t, N, S, p, a, b, r, k)
}
\arguments{
\item{t}{The time we want the of the option}

\item{N}{Instants of time the strategy lasts}

\item{S}{S=(S_0,...,S_t) vector of the prices at time t}

\item{p}{Up probability}

\item{a}{Down Interest}

\item{b}{Up interest}

\item{r}{Risk-free asset interest}

\item{k}{Strike price}
}
\value{
The price of the Asian Put Option
}
\description{
EPutOption_pricingfunction_det
Give the price of an Asian Put Option - Deterministic Formula p. 46
}
