% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot State Map with Accident Locations}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric or integer specifying the number for which the
is to be plotted.}

\item{year}{A numeric, integer, character or factor that should specify the year.
If multiple filenames are to be provided, input needs to be provided as vector.}
}
\value{
A plot of the state map for the specified state with all accident locations.
}
\description{
The function plots the map of a specified state (``state.num``) with all its
accident locations of a specified ``year``. Invalid ``state.num`` will throw
an error. Nothing will be plotted or returned if a valid state does not have
any observations.
}
\examples{
\dontrun{
fars_map_state(34, 2013)
}


}
