% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string specifying the file path.}
}
\value{
A tibble data.frame.
}
\description{
The functions can take the path to a csv file or a compressed csv file as input
and returns it as a tibble. Messages and progress status are not printed.
Execution stops if the specified file does not exist.
}
\examples{
\dontrun{
fars_read("data/accident_2013.csv.bz2")
}

}
