% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read compressed accident csv files of multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector specifying the years of the accident csv files.}
}
\value{
A list of tibbles with the columns ``MONTH`` and ``year``.
}
\description{
This function takes a vector as input that should specify the years of the
accident files. For every specified year, the filename is made, the file
is read and a tibble is created with only the columns ``MONTH`` and ``year``.
For each year that does not match an error is thrown. Furthermore the
function can only read data that sits in the current working directory.
}
\examples{
\dontrun{
fars_read_years(2014)
fars_read_years("2014")
fars_read_years(c(2013,2014))
}


}
