% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count monthly observations in accident data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector specifying the years of the accident csv files.}
}
\value{
A tibble data.frame with one row per month  and one column per year.
}
\description{
This function uses the function ``fars_read_years()`` to read all accident
csv files specified via parameter ``years`` and counts the number of observations
in each month.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014))
}

}
