% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill NA with the last no-NA value}
\usage{
fill_na(data, ...)
}
\arguments{
\item{data}{a data.frame or a vector}

\item{...}{col names character vector or just bare name.}
}
\value{
data.frame or a vector
}
\description{
See more: \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}}.
}
\examples{
x <- c(1, NA, NA, 2, 3)
y <- c(NA, NA, 1, NA, 3)
xy <- data.frame(x,y)
fill_na(x)
fill_na(y)
fill_na(xy, x, y)
fill_na(xy, c("x","y"))
fill_na(xy, starts_with("x"))
}
