% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-internals.R
\name{checkMissing}
\alias{checkMissing}
\alias{checkWeights}
\alias{checkID}
\alias{checkOffset}
\alias{getVarNames}
\title{Some tools for convert the date into c ++ files.}
\usage{
checkMissing(x, y)

checkWeights(w, n)

checkID(id)

checkOffset(o, y)

getVarNames(x)
}
\arguments{
\item{x}{a dateset}

\item{y}{a vector}

\item{w}{a vector of weight}

\item{n}{a size}

\item{id}{interaction.depth  which must less then 50}

\item{o}{the offset}
}
\value{
a vector of weight with length n

a vector of offset with y's length

a vector of names
}
\description{
A function to check the length vector
}
