% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.more.R
\name{gbm.more}
\alias{gbm.more}
\title{Adds additional trees with settled shrinkage and interaction.depth to a gbm object, .}
\usage{
gbm.more(
  object,
  n.new.trees = 100,
  interaction.depth = NULL,
  shrinkage = NULL,
  data = NULL,
  weights = NULL,
  offset = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A object created from an initial call
.}

\item{n.new.trees}{Integer specifying the number of additional trees to add
to \code{object}. Default is 100.}

\item{interaction.depth}{The maximum depth of variable interactions. A value
of 1 implies an additive model, a value of 2 implies a model with up to 2-way
interactions, etc. Default is objects's interaction.depth.}

\item{shrinkage}{The shrinkage parameter applied to each tree in the
expansion. Also known as the learning rate or step-size reduction; 0.001 to
0.1 usually work, but a smaller learning rate typically requires more trees.
Default is objects shrinkage}

\item{data}{An optional data frame containing the variables in the model. By
default the variables are taken from \code{environment(formula)}, typically
the environment from which \code{gbm} is called. If \code{keep.data=TRUE} in
the initial call to \code{gbm} then \code{gbm} stores a copy with the
object. If \code{keep.data=FALSE} then subsequent calls
must resupply the same dataset. It becomes the user's
responsibility to resupply the same data at this point.}

\item{weights}{An optional vector of weights to be used in the fitting
process. Must be positive but do not need to be normalized. If
\code{keep.data=FALSE} in the initial call to \code{gbm} then it is the
user's responsibility to resupply the weights.}

\item{offset}{A vector of offset values.}

\item{verbose}{Logical indicating whether or not to print out progress and
performance indicators (\code{TRUE}). If this option is left unspecified for
\code{gbm.more}, then it uses \code{verbose} from \code{object}. Default is
\code{FALSE}.}
}
\value{
A object.
}
\description{
Adds additional trees with settled shrinkage and interaction.depth to a gbm object, .
}
