% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date_breaks}
\alias{get_date_breaks}
\title{determine dates break
if the dates range is less than threshold then result is 1 week, otherwise 1 month}
\usage{
get_date_breaks(dates, threshold = 160)
}
\arguments{
\item{dates}{\code{Date} vector}

\item{threshold}{\code{numeric} threshold for week month choice}
}
\description{
determine dates break
if the dates range is less than threshold then result is 1 week, otherwise 1 month
}
