% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet a user}
\usage{
greet(name = "User")
}
\arguments{
\item{name}{character; name of user}
}
\value{
nothing! Just prints a greeting
}
\description{
This function prints a greeting.
}
\examples{
#Default use
greet()

#supply user name
greet("Eric")
}
