% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic_train}
\alias{logistic_train}
\title{Logistic Regression training}
\usage{
logistic_train(X, Y, testX, testY)
}
\arguments{
\item{X}{train set input}

\item{Y}{train set output}

\item{testX}{test set input}

\item{testY}{test set output}
}
\description{
The function will training the model with penalty term 1 to 200 to find the best fit
It will output a graph states the accuracy at given regularization term
}
