% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create filename of a compressed csv-file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric, integer, character or factor that should specify the year.
If multiple filenames are to be provided, input needs to be provided as vector.}
}
\value{
A character string representing the filename of a compressed accident
  csv for the specified year.
}
\description{
The functions takes an input string that should specify the year and
returns the filename of the accident data for the specified year.
}
\examples{
\dontrun{
make_filename(2013)
}

}
