% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{make_shades}
\alias{make_shades}
\title{Make shades}
\usage{
make_shades(colour, n, lighter = TRUE)
}
\arguments{
\item{colour}{The colour to make shades of}

\item{n}{The number of shades to make, at least 1}

\item{lighter}{Whether to make lighter (\code{TRUE}) or darker (\code{FALSE})
shades}
}
\value{
A vector of \code{n} colour hex codes
}
\description{
Given a colour make \code{n} lighter or darker shades
}
\examples{
# Five lighter shades
make_shades("goldenrod", 5)
# Five darker shades
make_shades("goldenrod", 5, lighter = FALSE)
}
