% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg-package.R
\docType{package}
\name{mypkg-package}
\alias{mypkg}
\alias{mypkg-package}
\title{mypkg: My Personal Package}
\description{
This is my personal package. It contains some handy functions that
    I find useful for my projects.
}
\author{
\strong{Maintainer}: Package Creator \email{package.creator@mypkg.com}

Other contributors:
\itemize{
  \item Package Contributor \email{package.contributor@mypkg.com} [contributor]
}

}
\keyword{internal}
