% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg.R
\docType{package}
\name{mypkg}
\alias{mypkg}
\title{Mypkg lead file}
\description{
This package contains 3 training model: GBM, logistic, and SVM:1.Gradient Boosting Regression(GMB) is a learning method that combines many decision trees as the weak predictors. into an ensemble, forming a strong predictor. In this package, we are focus on the Bernoulli distribution, and write a series of tools to improve its accuracy and reduce its running time. Otherwise, we are trying to use control weight of each observations to anti the unbalance dataset. 2.logistic resgression takes both training and validation dataset as parameter and recursively finding best fit penalty term to reuce the influence of extreme values.3.A SVM package implemted by Python with SMO algorithm  method.
}
\author{
Henglin Huang, Jing Xie , Hujie Han, Fang yi Wang
}
\keyword{package}
