% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPkg_fun.R
\name{myPkg_fun}
\alias{myPkg_fun}
\title{Sample function}
\usage{
myPkg_fun(path = ".")
}
\arguments{
\item{path}{The path where the RDS file is saved. Defaults to local path.}
}
\value{
`NULL`
}
\description{
Sample function saves "Hello World" in RDS file
}
