% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-json.R
\name{parse_json}
\alias{parse_json}
\title{Title}
\usage{
parse_json(file)
}
\arguments{
\item{file}{}
}
\description{
Title
}
\examples{
\dontrun{
write_csv(parse_json("data/3-1.json"),"data/parse-3-1.csv")
write_csv(parse_json("data/3-2.json"),"data/parse-3-2.csv")
write_csv(parse_json("data/3-3.json"),"data/parse-3-3.csv")
}
}
