% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{plot_colours}
\alias{plot_colours}
\title{Plot colours}
\usage{
plot_colours(colours)
}
\arguments{
\item{colours}{Vector of colour to plot}
}
\value{
A ggplot2 object
}
\description{
Plot a vector of colours to see what they look like
}
\examples{
shades <- make_shades("goldenrod", 5)
plot_colours(shades)
}
