% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png2svg.R
\name{png2svg}
\alias{png2svg}
\title{Convert png to svg}
\usage{
png2svg(png_path, svg_path, width = 800, height = 800)
}
\arguments{
\item{png_path}{path to the png file,character.}

\item{svg_path}{out path of the svg file, character.}

\item{width}{numeric, e.g: 200}

\item{height}{numeric, e.g: 200}
}
\value{
a svg file
}
\description{
Convert png to svg
}
\details{
The quality of svg file depends on the png file.
}
\examples{
png_path <- system.file("img", "Rlogo.png", package="png")
svg_path <- tempfile(fileext = ".svg")
png2svg(png_path, svg_path)
file.show(svg_path)
png2svg(png_path, svg_path, 200, 200)

}
