% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{process covid data
Compute additional variables}
\usage{
process_data(data, k = 7)
}
\arguments{
\item{data}{\code{data.frame}, covid19 data.}

\item{k}{\code{integer}, window size for \code{roll_mean}}
}
\value{
data.frame
}
\description{
process covid data
Compute additional variables
}
\details{
it computes actives and daily figures
}
\examples{
df <- read_data("case_time_series.csv","covid_data")
process_data(df)
}
\seealso{
\code{\link[=roll_mean]{roll_mean()}}
}
