% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnmbrm.r
\name{Qnmbrm}
\alias{Qnmbrm}
\title{Quick Nodes Morph Boosted Regression Modeling (QNMBRM)}
\usage{
Qnmbrm(
  formula = formula(data),
  data = list(),
  weights = NULL,
  stage.level = 3,
  stage.rate = 2,
  n.trees = 100,
  interaction.depth = NULL,
  n.minobsinnode = 10,
  shrinkage = 0.1,
  shrinkage.top = 0.1
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. The formula
may include an offset term (e.g. y~offset(n)+x). If
\code{keep.data = FALSE} in the initial call to \code{gbm} then it is the
user's responsibility to resupply the offset to gbm.more.}

\item{data}{an optional data frame containing the variables in the model. By
default the variables are taken from \code{environment(formula)}, typically
the environment from which \code{gbm} is called. If \code{keep.data=TRUE} in
the initial call to \code{gbm} then \code{gbm} stores a copy with the
object. If \code{keep.data=FALSE} then subsequent calls to
gbm.more must resupply the same dataset. It becomes the user's
responsibility to resupply the same data at this point.}

\item{weights}{A vector of weights of the same length as the \code{y}.}

\item{stage.level}{integer the levels of stages groups with different weight and
study rate. Default is 3.}

\item{stage.rate}{integer the rate of difference betweent differents levels.
it should be a number larger then 1. this effect the size and study rate of
most of stages. Default is 2}

\item{n.trees}{Integer specifying the total number of trees to fit. This is
equivalent to the number of iterations and the number of basis functions in
the additive expansion. Default is 100.}

\item{interaction.depth}{Integer specifying the maximum depth of each tree
(i.e., the highest level of variable interactions allowed). A value of 1
implies an additive model, a value of 2 implies a model with up to 2-way
interactions, etc. Default is automatic fill a considered number.}

\item{n.minobsinnode}{Integer specifying the minimum number of observations
in the terminal nodes of the trees. Note that this is the actual number of
observations, not the total weight.}

\item{shrinkage}{a shrinkage parameter applied to each tree in the
expansion. Also known as the learning rate or step-size reduction; 0.001 to
0.1 usually work, but a smaller learning rate typically requires more trees.
Default is 0.1.}

\item{shrinkage.top}{a shrinkage parameter start to working with.}
}
\value{
A gbm object.
}
\description{
adpet the weight to reduce the FLOPS of algorithm
}
\author{
Henglin Huang
}
