% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data}
\alias{read_data}
\title{read covid data
Function to read covid data from case_time_series.csv}
\usage{
read_data(filename, path, sep = ",")
}
\arguments{
\item{filename}{\code{character}, file name.}

\item{path}{\code{character}, folder name.}

\item{sep}{\code{character}, separator for \code{read.csv}, default = ','}
}
\value{
data.frame
}
\description{
read covid data
Function to read covid data from case_time_series.csv
}
\examples{
read_data("case_time_series.csv","covid_data")

}
\seealso{
\code{\link[utils:read.table]{utils::read.csv()}}
}
