% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_axes}
\alias{scale_axes}
\title{Scale plot axes
defined plot style for axes}
\usage{
scale_axes(pp, datebreaks = "1 week")
}
\arguments{
\item{pp}{\code{ggplot} object}

\item{datebreaks}{\code{character} x axis date breaks, default = "1 week"}
}
\description{
Scale plot axes
defined plot style for axes
}
\seealso{
\code{\link[=scale_x_date]{scale_x_date()}} \code{\link[=scale_y_continuous]{scale_y_continuous()}}
}
