\name{sumC}
\alias{sumC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate mean.
}
\description{
This function will calculate the mean of a variable
}
\usage{
sumC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Any numeric value
}
}
\details{
To calculate mean.
}
\value{
This will calculate mean
}
\author{
Jahangir Alam
}