\name{sumR}
\alias{sumR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate summation using R
}
\description{
This function takes input as numeric vector and calculate sum using R only
}
\usage{
sumR(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
 A vector of numeric values
}
}
\details{
To calculate sum of numeric vector using R
}
\value{
Return mean of input vector \code{x}
}
\author{
Jaynal abedin
}
