% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmWrapper.R
\name{svmWrapper}
\alias{svmWrapper}
\title{Title wrapper function for SVM}
\usage{
svmWrapper(X, y, kernal_type, size)
}
\arguments{
\item{X}{is the training data and}

\item{y}{is the label of taining data}

\item{kernal_type}{is the type of different kernel function}

\item{size}{is the size of train test ratio}
}
\value{
the accruracy of svm model
}
\description{
Title wrapper function for SVM
}
\examples{
svmWrapper(X, y, kernal_type, size)
}
