\name{test.c}
\alias{test.c}
\title{mypkg test.c}
\description{
  mypkg test.c for 'RcallC.R'.
}
\usage{
  test.c(a)
}
\arguments{
  \item{a}{a = 1 : 9.}
  }

\value{
  \item{a}{A test input.}
}

\details{
  This is a simple example 'test.c' in 'mypkg'. 
}

\examples{

a<-1:9
test.c(a)

}

\author{MyLastName, MyFirstName <my@email.address>}

\references{
  \url{http://my.web.page.address}  
}

\keyword{misc}
