\name{test.f}
\alias{test.f}
\title{mypkg test.f}
\description{
  mypkg test.f for 'RcallF.R'.
}
\usage{
  test.f(a)
}
\arguments{
  \item{a}{a = 1 : 9.}
  }

\value{
  \item{a}{A test input.}
}

\details{
  This is a simple example 'test.f' in 'mypkg'. 
}

\examples{

a<-1:9
test.f(a)

}

\author{MyLastName, MyFirstName <my@email.address>}

\references{
  \url{http://my.web.page.address}  
}

\keyword{misc}
