% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_plot}
\alias{theme_plot}
\title{theme plot
defined plot style}
\usage{
theme_plot()
}
\description{
theme plot
defined plot style
}
\examples{
library(ggplot2)
p1 <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  labs(title = "Fuel economy declines as weight increases") +
  theme_plot()
}
