% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Time series plot}
\usage{
ts_plot(d)
}
\arguments{
\item{d}{data frame with two columns}
}
\value{
interactive dygraph plot
}
\description{
Plot time series interactively.
}
\examples{
ts_plot(df_exp)
}
