\name{varC}
\alias{varC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate variance
}
\description{
This function will calculate the variance of a vector
}
\usage{
varC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x is any numeric vector
}
}
\details{
This function will calculate the variance
}
\value{
Any numeric value
}

\author{
Jahangir Alam
}
