% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaOptions_Deterministic.R
\name{VCallOption_hedgingfunction_det}
\alias{VCallOption_hedgingfunction_det}
\title{VCallOption_hedgingfunction_det
Calculate the portfolio allocation hedging the European Call Option - Formula p. 47 Proposition 3.2}
\usage{
VCallOption_hedgingfunction_det(t, N, S, p, a, b, r, k, pi_0)
}
\arguments{
\item{t}{The time we want the of the option}

\item{N}{Instants of time the strategy lasts}

\item{S}{S = S_(t-1)}

\item{p}{Up probability}

\item{a}{Down Interest}

\item{b}{Up interest}

\item{r}{Risk-free asset interest}

\item{k}{Strike price}

\item{pi_0}{Price of Risk-less assset at time t=0}
}
\value{
Eps and Eta rispectively allocation of Risk-asset and Free-risk asset
}
\description{
VCallOption_hedgingfunction_det
Calculate the portfolio allocation hedging the European Call Option - Formula p. 47 Proposition 3.2
}
