% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaOptions_Deterministic.R
\name{VCallOption_pricingfunction_det}
\alias{VCallOption_pricingfunction_det}
\title{VPutOption_pricingfunction_det
Give the price of an European Call Option - Deterministic Formula p. 46}
\usage{
VCallOption_pricingfunction_det(t, N, S, p, a, b, r, k)
}
\arguments{
\item{t}{The time we want the of the option}

\item{N}{Instants of time the strategy lasts}

\item{S}{S_t, price of the risk asset at time t}

\item{p}{Up probability}

\item{a}{Down Interest}

\item{b}{Up interest}

\item{r}{Risk-free asset interest}

\item{k}{Strike price}
}
\value{
The price of the European Call Option
}
\description{
VPutOption_pricingfunction_det
Give the price of an European Call Option - Deterministic Formula p. 46
}
