% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaOptions_Random.R
\name{VCallOption_pricingfunction_rand}
\alias{VCallOption_pricingfunction_rand}
\title{Title  VCAllOption_pricingfunction_rand
Give the price of an European Call Option - Montecarlo Formula p. 46}
\usage{
VCallOption_pricingfunction_rand(t, N, S, p, a, b, r, k, M)
}
\arguments{
\item{t}{The time we want the of the option}

\item{N}{Instants of time the strategy lasts}

\item{S}{S_t, price of the risk asset at time t}

\item{p}{Up probability}

\item{a}{Down Interest}

\item{b}{Up interest}

\item{r}{Risk-free asset interest}

\item{k}{Strike price}

\item{M}{Number of independent variables in Montecarlo}
}
\value{
The price of the European Call Option
}
\description{
Title  VCAllOption_pricingfunction_rand
Give the price of an European Call Option - Montecarlo Formula p. 46
}
