% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightCal.R
\name{weightCal}
\alias{weightCal}
\title{WeightCAL}
\usage{
weightCal(dataframe, remove, cate, patten = "a")
}
\arguments{
\item{dataframe}{the data to offer a set of date to find the weight}

\item{remove}{a vectors of the name of the variable move away from the weight
calculate}

\item{cate}{a vector of the name of variable should treat as vector but its
dataset is integer}

\item{patten}{the method to calculate the each weight, when patten equals "a"
we add all wight togetherm, when patten equals "m", we mutiply them together.
when patten equal "l", we mutiple the logs of their weight}
}
\value{
A vector of the weight
}
\description{
A tools to calculate the weight of each value in a dataset
}
\author{
Henglin Huang
}
