% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{This is data set I created}
\format{\preformatted{ num [1:10] 0.2572 -0.5297 0.0872 0.499 1.1313 ...
}}
\usage{
x
}
\description{
It is a stupid data set.
}
\keyword{datasets}

