% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baked_goods}
\alias{baked_goods}
\title{Cupcake and Muffin Ingredients}
\format{
\subsection{\code{baked_goods}}{

A data frame with 870 rows and 4 columns:
\describe{
\item{type}{Muffin or cupcake}
\item{recipe_id}{Unique numeric identifier for recipe}
\item{ingredient}{Ingredient}
\item{amount_cups}{Amount of ingredient in cups}
...
}
}
}
\source{
\url{https://raw.githubusercontent.com/cct-datascience/sortee-rpkg/main/baked_goods_long.csv}
}
\usage{
baked_goods
}
\description{
Recipe info from allrecipes for many delicious cupcakes and muffins
}
\keyword{datasets}
