% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet a user}
\usage{
greet(name = "User")
}
\arguments{
\item{name}{character; name of user}
}
\value{
nothing! just prints a greeting
}
\description{
This function greets a user if you tell it their name.
}
\examples{
#Default use
greet()

#Supply user name
greet("Annie")

}
