\name{meanX}
\alias{meanX}
Calculate mean
\title{
 Calculate mean
}
\description{
Calculate mean
}
\usage{
meanX(x)
}
Calculate mean
\arguments{
  \item{x}{
Calculate mean
}
}
\details{
Calculate mean
}
\value{
Calculate mean
}
\references{
Calculate mean
}
\author{
sonia
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
    n <- length(x)
    meanX <- sum(x)/n
    return(meanX)
  }
}

\keyword{ Calculate mean }
\keyword{ Calculate mean}% __ONLY ONE__ keyword per line
