% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Multivariate normal distribution density}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix, with \code{n} columns (the observations) and \code{p} rows}

\item{mean}{a vector of means}

\item{varcovM}{a variance-covariance matrix}

\item{Log}{a logical parameter, with default value to \code{TRUE}}
}
\value{
a list containing the matrix \code{x}, and a vector of length \code{n}
of the multivariate normal distribution density values at those points.
}
\description{
Function to compute the value of the density of a multivariate normal distribution.
}
\examples{
matrice <- cbind(c(1,2), c(3,4))
mean <- c(1, 1)
varcov <- diag(1, 2)
mvnpdf(x = matrice, mean = mean, varcovM = varcov)
}
