\name{mypkg-package}
\alias{mypkg-package}
\alias{mypkg}
\docType{package}
\title{
What the package does (short line)
Calculate mean
}
\description{
More about what it does (maybe more than one line)
Calculate mean
}
\details{
\tabular{ll}{
Package: \tab mypkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-08\cr
License: \tab What license is it under?\cr
}
Calculate mean
}
\author{
sonia

Maintainer:sonia <soniruatat10@gmail.com>
Calculate mean
}
\references{
Calculate mean
}


\keyword{ package }
\seealso{

}
\examples{

}
