% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg-package.R
\docType{package}
\name{mypkg-package}
\alias{mypkg}
\alias{mypkg-package}
\title{mypkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/melaniehuchon/mypkg}
  \item Report bugs at \url{https://github.com/melaniehuchon/mypkg/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
