% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg-package.R
\docType{package}
\name{mypkg-package}
\alias{mypkg}
\alias{mypkg-package}
\title{mypkg: Explore Your R libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Mutaz94/mypkg}
  \item Report bugs at \url{https://github.com/Mutaz94/mypkg/issues}
}

}
\author{
\strong{Maintainer}: Mutaz Jaber \email{jaber038@umn.edu} (\href{https://orcid.org/0000-0002-7536-8753}{ORCID})

}
\keyword{internal}
