\name{mypkg-package}
\alias{mypkg-package}
\alias{mypkg}
\docType{package}
\title{
\packageTitle{mypkg}
}
\description{
\packageDescription{mypkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mypkg}
\packageIndices{mypkg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mypkg}

Maintainer: \packageMaintainer{mypkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
