% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg-package.R
\docType{package}
\name{mypkg-package}
\alias{mypkg}
\alias{mypkg-package}
\title{mypkg: Kidney ST Plotting package}
\description{
Here lives the function I use to plot kidney ST.
}
\author{
\strong{Maintainer}: Simon Mo \email{c.mo@wustl.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
