\name{sumC}
\alias{sumC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate sum using C
}
\description{
To calculate summation of a numeric vector using C code
}
\usage{
sumC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
 A vector of numeric values
}
}
\details{
Calculating sum of a vector using C loop
}
\value{
Mean of input vector \code{x}
}
\author{
Jaynal Abedin
}

